misc fixes, add documentations, optimize webpack build
