+++ package.json
@@ -2,7 +2,7 @@
     "name": "@springrole/springwallet",
     "version": "0.1.6",
     "description": "Wallet for SpringRole users",
-    "main": "dist/springwallet.js",
+    "main": "src/springwallet.js",
     "engines": {
         "node": ">=8.0.0"
     },
@@ -40,7 +40,6 @@
     },
     "homepage": "https://github.com/SpringRole/SpringWallet#readme",
     "dependencies": {
-        "babel-eslint": "^10.0.3",
         "bip39": "^3.0.2",
         "ethers": "^4.0.33",
         "sweetalert2": "^8.15.2"
@@ -53,8 +52,8 @@
         "@babel/plugin-transform-runtime": "^7.6.0",
         "@babel/preset-env": "^7.6.0",
         "@babel/runtime": "^7.6.0",
+        "babel-eslint": "^10.0.3",
         "babel-loader": "^8.0.6",
-        "babel-register": "^6.26.0",
         "eslint": "^6.3.0",
         "eslint-config-prettier": "^6.2.0",
         "eslint-config-standard": "^14.1.0",
@@ -64,8 +63,6 @@
         "eslint-plugin-promise": "^4.2.1",
         "eslint-plugin-standard": "^4.0.1",
         "prettier": "^1.18.2",
-        "rollup": "^1.21.2",
-        "rollup-plugin-babel": "^4.3.3",
         "webpack": "^4.39.1",
         "webpack-cli": "^3.3.6"
     }
